Add new commenting feature
A list of comments can be added to the repo config, and workspaces can then reference them.
The bot will then post that comment if the PR matches the workspace.

This can be used to explain to reviewers what is expected from them and how to trigger a speculative plan.